Initial commit
This is already sort of working, the biggest problem being the
slowness of the WARC. Next step for me will be to parallelize the crap
out of this - even with the interpretor lock this should give me a big
boost, because I'm mostly waiting for packets.

There's more TODOs, including handling screenshot pages and finding a
way to crawl the entire archive page (for some reason I can only see
the first ~80 of them).